requirements.txt
django==4.2.2
Pillow==9.5.0
djangorestframework==3.14.0
django-rest-swagger==2.2.0